+++ gulpfile.js
@@ -1,6 +1,5 @@
 var { src, dest } = require('gulp');
 var gulp = require('gulp');
-connect = require('gulp-connect');
 var include = require('gulp-include');
 var concat = require('gulp-concat');
 var uglify = require("gulp-uglify")
@@ -60,7 +59,6 @@ gulp.task('html', function() {
     return src('*.html')
       .pipe(concat('app.html'))
       .pipe(dest('app/'))
-      .pipe(connect.reload());
 });
 
 gulp.task('watch', function(){
@@ -88,21 +86,8 @@ gulp.task('browser-sync', function() {
     });
 });
 
-gulp.task('webserver', function() {
-    connect.server({
-        root: 'app',
-        index: "app.html",
-    });
-    console.log("----------------------------------");
-    console.log("[\x1b[31m v0.5 Catalyst App \x1b[37m] \- \x1b[32mdeveloped by Daniel Kandilas");
-    console.log("----------------------------------");
-});
-
 //local build
 gulp.task('local', gulp.parallel('js', 'scss', 'themes', 'html', 'watch','browser-sync'));
 
 //deployment
 gulp.task('prod', gulp.parallel('js', 'scss', 'themes', 'html'));
-
-//server
-//gulp.task('server', gulp.series('webserver'));
